test file structure + test for get spouses&childre
