+++ mern/server/testing/helper.js
@@ -54,6 +54,121 @@ class TestDbHelper {
       .insertOne(document);
     return ops[0];
   }
+  async  createSampleUsers() {
+    const user1 = await this.createDoc('users', {
+      name: "First user",
+      dateOfBirth: "010101",
+      address: "Lygten 1",
+      CPR: "0101010001",
+      maritalStatusId: "8",
+      spouse: { _id: 123456, name: "some guy" },
+      genderIdentification: "0001",
+    });
+    const user2 = await this.createDoc('users', {
+      name: "Second user",
+      dateOfBirth: "020280",
+      CVR: "12345678",
+      address: "Lygten 3",
+      companyNamy: "EG",
+      CPR: "0202800002",
+      genderIdentification: "0001",
+    });
+    const user3 = await this.createDoc('users', {
+      name: "Third user",
+      dateOfBirth: "101245",
+      address: "Lygten",
+      maritalStatusId: "2",
+      CPR: "1012450002",
+      spouse: { _id: 123456, name: "some guy" },
+      genderIdentification: "0002",
+    });
+    const user4 = await this.createDoc('users', {
+      name: "Fourth user",
+      dateOfBirth: "101215",
+      address: "Lygten",
+      CPR: "1012150002",
+      maritalStatusId: "8",
+      genderIdentification: "0002",
+    });
+    const user5 = await this.createDoc('users', {
+      name: "Fifth user",
+      dateOfBirth: "141286",
+      address: "Lygten",
+      maritalStatusId: "1",
+      CPR: "0202840001",
+      genderIdentification: "0001",
+    });
+    const user6 = await this.createDoc('users', {
+      name: "Sixth user",
+      dateOfBirth: "181245",
+      address: "Lygten",
+      maritalStatusId: "3",
+      CPR: "1012450002",
+      genderIdentification: "0002",
+    });
+    const user7 = await this.createDoc('users', {
+      name: "Seventh user",
+      dateOfBirth: "020280",
+      address: "Lygten 3",
+      maritalStatusId: "4",
+      CPR: "0202800001",
+      genderIdentification: "0001",
+    });
+    const user8 = await this.createDoc('users', {
+      name: "Eights user",
+      dateOfBirth: "020280",
+      address: "Lygten 3",
+      maritalStatusId: "5",
+      CPR: "0202800001",
+      spouse: { _id: 123456, name: "some guy" },
+      genderIdentification: "0001",
+    });
+    const user9 = await this.createDoc('users', {
+      name: "Nineth user",
+      dateOfBirth: "101245",
+      address: "Lygten",
+      maritalStatusId: "6",
+      CPR: "1012450002",
+      genderIdentification: "0002",
+    });
+    const user10 = await this.createDoc('users', {
+      name: "Tenth user",
+      dateOfBirth: "101255",
+      address: "Lygten",
+      CPR: "1012150002",
+      maritalStatusId: "7",
+      spouse: { _id: 123456, name: "some guy" },
+      genderIdentification: "0002",
+    });
+  
+    const user11 = await this.createDoc('users', {
+      name: "Eleventh user",
+      dateOfBirth: "101255",
+      address: "Lygten",
+      CPR: "1012150002",
+      maritalStatusId: "7",
+      spouse: { _id: 123456, name: "some guy" },
+      genderIdentification: "0002",
+    });
+  
+    return {
+      user1,
+      user2,
+      user3,
+      user4,
+      user5,
+      user6,
+      user7,
+      user8,
+      user9,
+      user10,
+      user11,
+    };
+
+  }
 }
+
+
+
 
 module.exports = TestDbHelper
